fix railstie spec
The original spec did a bunch of voodoo and broke with the fix for
adding the Pry::CLI option parsing.

The fix is to run all tests through Rails::Console.start and check the
output of pry specific command.

Note: the $stdout = StringIO.new is there to prevent a bunch of text
from showing up in the test run.